백준 20551번 2트 실패 코드

+++ Question/src/Sort/Q20551.java
@@ -11,14 +11,14 @@ public class Q20551 {
         int N = Integer.parseInt(st.nextToken());
         int M = Integer.parseInt(st.nextToken());
 
-        int[] arr = new int[N];
-        for(int i=0; i<N; i++) arr[i] = Integer.parseInt(br.readLine());
+        long[] arr = new long[N];
+        for(int i=0; i<N; i++) arr[i] = Long.parseLong(br.readLine());
         Arrays.sort(arr);
 
         StringBuilder sb = new StringBuilder();
 
         for(int i=0; i<M; i++){
-            int a = Integer.parseInt(br.readLine());
+            long a = Long.parseLong(br.readLine());
             int tmp = -1;
             for(int j=0; j<N; j++){
                 if(a==arr[j]){
